adds test for deployment
fixes errors that where not raised
