+++ ci/infra/testrunner/kubectl/kubectl.py
@@ -1,5 +1,7 @@
 import os
 
+from timeout_decorator import timeout
+
 from platforms.platform import Platform
 from utils.format import Format
 from utils.utils import (step, Utils)
@@ -10,24 +12,15 @@ class Kubectl:
         self.conf = conf
         self.utils = Utils(self.conf)
         self.platform = Platform.get_platform(conf, platform)
-        self.cwd = "{}/test-cluster".format(self.conf.workspace)
-
-
-    def get_pods(self):
-        print("gets pods using kubectl")
-        try:
-            self._run_kubectl("get po -o wide")
-        except Exception as ex:
-            print("Received the following error {}".format(ex))
 
 
     def create_deployment(self, name, image):
         print("create a new deployment {}".format(name))
         try:
             self._run_kubectl("create deployment {name}  --image={image}"
-                             .format(name=name, image=image))
+                                .format(name=name, image=image))
         except Exception as ex:
-            print("Received the following error {}".format(ex))
+            raise Exception("Error executing cmd {}") from ex
 
 
     def scale_deployment(self, name, replicas):
@@ -36,7 +29,7 @@ class Kubectl:
             self._run_kubectl("scale deployment {name} --replicas={replicas}"
                              .format(name=name, replicas=replicas))
         except Exception as ex:
-            print("Received the following error {}".format(ex))
+            raise Exception("Error executing cmd {}") from ex
 
 
     def expose_deployment(self, name, port, nodeType="NodePort"):
@@ -45,7 +38,7 @@ class Kubectl:
             self._run_kubectl("expose deployment {name} --port={port} --type={nodeType}"
                              .format(name=name, port=port, nodeType=nodeType))
         except Exception as ex:
-            print("Received the following error {}".format(ex))
+            raise Exception("Error executing cmd {}") from ex
 
 
     def wait_deployment(self, name, timeout):
@@ -54,7 +47,7 @@ class Kubectl:
             self._run_kubectl("wait --for=condition=available deploy/{name} --timeout={timeout}m"
                              .format(name=name, timeout=timeout))
         except Exception as ex:
-            print("Received the following error {}".format(ex))
+            raise Exception("Error executing cmd {}") from ex
 
 
     def count_available_replicas(self, name):
@@ -64,8 +57,7 @@ class Kubectl:
                                        .format(name=name))
             return int(result)
         except Exception as ex:
-            print("Received the following error {}".format(ex))
-            return 0
+            raise Exception("Error executing cmd {}") from ex
 
 
     def get_service_port(self, name):
@@ -75,25 +67,24 @@ class Kubectl:
                                        .format(name=name))
             return int(result)
         except Exception as ex:
-            print("Received the following error {}".format(ex))
-            return 0
+            raise Exception("Error executing cmd {}") from ex
 
-    def test_service(self, name, path="/"):
+
+    def test_service(self, name, path="/", worker=0):
         print("curl deployment {}".format(name))
         ip_address = self.platform.get_nodes_ipaddrs("worker")
         port = self.get_service_port(name)
         try:
             return self.utils.runshellcommand_withoutput("curl {ip}:{port}{path}"
-                                                    .format(ip=ip_address[0],port=port,path=path), False)
+                                                        .format(ip=ip_address[worker],port=port,path=path), False)
         except Exception as ex:
-            print("Received the following error {}".format(ex))
-            return ""
+            raise Exception("Error executing cmd {}") from ex
 
 
     def _run_kubectl(self, command):
         try:
             return self.utils.runshellcommand_withoutput("kubectl --kubeconfig={cwd}/test-cluster/admin.conf -o json {command}"
-                                                  .format(command=command,cwd=self.conf.workspace), False)
+                                                        .format(command=command,cwd=self.conf.workspace), False)
         except Exception as ex:
-            print("Received the following error {}".format(ex))
-            return ""
+            raise Exception("Error executing cmd {}") from ex
+
